Fix: js > jsx
