+++ components/_layouts/main.layout.jsx
@@ -1,31 +1,29 @@
 import React, { ReactNode, useRef } from "react";
 import Link from "next/link";
 import Script from "next/script";
-import Navbar from "../defaults/Navbar";
-import Footer from "../defaults/Footer";
+import Navbar from "../_common/navbar";
+import Footer from "../_common/footer";
 // import { LocomotiveScrollProvider } from "react-locomotive-scroll";
 // import "locomotive-scroll/dist/locomotive-scroll.min.css";
-import Meta from "../defaults/Meta";
-import Head from "next/head";
+import Meta from "../_common/meta";
 
 const Layout = ({ children }) => {
     const containerRef = useRef(null);
     return (
         <>
-            <Head>
-                <Meta
-                    title="Bitzquad | Solutions Beyond Technology | Software Company"
-                    description="We are a team of thinkers, engineers, designers, and marketers who represent different cultures worldwide, working together to build better solutions
+            <Meta
+                title="Bitzquad | Solutions Beyond Technology | Software Company"
+                description="We are a team of thinkers, engineers, designers, and marketers who represent different cultures worldwide, working together to build better solutions
                             for a better world.  We strive to provide sustaining digital solutions in Information Systems, Business Process Re-engineering, Branding and Digital Marketing, and
                             E-Business services that serve our stakeholders' best interests."
-                    keywords="Bitzquad, Solutions Beyond Technology, Software Company Information Systems, Business Process Re-engineering, Branding, Digital Marketing, E-Business services"
-                    url={`${process.env.NEXT_PUBLIC_API_URL}/`}
-                    imagefb={`${process.env.NEXT_PUBLIC_API_URL}/og-img.webp`}
-                    alt="We are a team of thinkers, engineers, designers, and marketers who represent different cultures worldwide, working together to build better solutions
+                keywords="Bitzquad, Solutions Beyond Technology, Software Company Information Systems, Business Process Re-engineering, Branding, Digital Marketing, E-Business services"
+                url={`${process.env.NEXT_PUBLIC_API_URL}/`}
+                imagefb={`${process.env.NEXT_PUBLIC_API_URL}/og-img.webp`}
+                alt="We are a team of thinkers, engineers, designers, and marketers who represent different cultures worldwide, working together to build better solutions
                     for a better world.  We strive to provide sustaining digital solutions in Information Systems, Business Process Re-engineering, Branding and Digital Marketing, and
                     E-Business services that serve our stakeholders' best interests."
-                ></Meta>
-            </Head>
+            ></Meta>
+
             <Script src="/assets/js/cursorControls.js" />
             {/* <LocomotiveScrollProvider
                 innerRef={containerRef}
