+++ components/_layouts/subpage.layout.jsx
@@ -2,8 +2,8 @@ import React, { ReactNode, useRef } from "react";
 import Link from "next/link";
 import Script from "next/script";
 import Head from "next/head";
-import Navbar from "../defaults/Navbar";
-import Footer from "../defaults/Footer";
+import Navbar from "../_common/navbar";
+import Footer from "../_common/footer";
 // import { LocomotiveScrollProvider } from "react-locomotive-scroll";
 // import "locomotive-scroll/dist/locomotive-scroll.min.css";
