+++ components/index.js
@@ -1,23 +1,23 @@
 // layouts
-export { default as LayoutMain } from "./layouts/LayoutMain";
-export { default as LayoutSubPages } from "./layouts/LayoutSubPages";
+export { default as LayoutMain } from "./_layouts/main.layout";
+export { default as LayoutSubPages } from "./_layouts/subpage.layout";
 
 // defaults
-export { default as Navbar } from "./defaults/Navbar";
-export { default as Footer } from "./defaults/Footer";
-export { default as Loader } from "./defaults/Loader";
-export { default as SideMenu } from "./defaults/SideMenu";
+export { default as Navbar } from "./_common/navbar";
+export { default as Footer } from "./_common/footer";
+export { default as Loader } from "./_common/loader";
+export { default as SideMenu } from "./_common/sidemenu";
 
 // elements
-export { default as Button } from "./elements/Button";
+export { default as Button } from "./elements/button";
 
 // Home
-export { default as Landing } from "./home/Landing";
-export { default as Testimonials } from "./home/Testimonials";
-export { default as Collaborate } from "./home/Collaborate";
-export { default as Services } from "./home/Services";
-export { default as OurTeam } from "./home/OurTeam";
-export { default as OurPartners } from "./home/OurPartners";
-export { default as ProjectSection } from "./home/ProjectSection";
+export { default as Landing } from "./home/landing";
+export { default as Testimonials } from "./home/testimonials";
+export { default as Collaborate } from "./home/collaborate";
+export { default as Services } from "./home/services";
+export { default as OurTeam } from "./home/ourteam";
+export { default as OurPartners } from "./home/ourpartners";
+export { default as ProjectSection } from "./home/projectsection";
 
-export { default as Code } from "./blog/Code";
+export { default as Code } from "./blog/code";
