--- constants/schemas/News.js
@@ -1,47 +1,0 @@
-import mongoose from "mongoose";
-
-const NewsSchema = new mongoose.Schema(
-    {
-        title: {
-            type: String,
-            required: true,
-        },
-        content: {
-            type: String,
-            required: true,
-        },
-        thumbnail: {
-            type: String,
-            required: true,
-        },
-        category: {
-            type: String,
-            required: true,
-        },
-        writer: {
-            name: {
-                type: String,
-                required: true,
-            },
-            title: {
-                type: String,
-                required: true,
-            },
-            thumbnail: {
-                type: String,
-                required: true,
-            },
-        },
-        status: {
-            type: String,
-            default: "Pending",
-        },
-        deleted: {
-            type: Boolean,
-            default: false,
-        },
-    },
-    { timestamps: true }
-);
-
-export default mongoose.models.News || mongoose.model("News", NewsSchema);
