--- constants/schemas/contact.js
@@ -1,33 +1,0 @@
-import mongoose from "mongoose";
-
-const ContactSchema = new mongoose.Schema(
-    {
-        name: {
-            type: String,
-            required: true,
-        },
-        email: {
-            type: String,
-            required: true,
-        },
-        mobile: {
-            type: String,
-            default: null,
-        },
-        title: {
-            type: String,
-            default: null,
-        },
-        message: {
-            type: String,
-            required: true,
-        },
-        deleted: {
-            type: Boolean,
-            default: false,
-        },
-    },
-    { timestamps: true }
-);
-
-export default mongoose.models.Contact || mongoose.model("Contact", ContactSchema);
