--- constants/schemas/newsletter.js
@@ -1,21 +1,0 @@
-import mongoose from "mongoose";
-
-const NewsletterSchema = new mongoose.Schema(
-    {
-        email: {
-            type: String,
-            required: true,
-        },
-        subscribed: {
-            type: Boolean,
-            default: true,
-        },
-        deleted: {
-            type: Boolean,
-            default: false,
-        },
-    },
-    { timestamps: true }
-);
-
-export default mongoose.models.Newsletter || mongoose.model("Newsletter", NewsletterSchema);
