+++ pages/projects/[id].jsx
